+++ core/storage/page_cache.rs
@@ -870,7 +870,7 @@ mod tests {
     }
 
     pub fn page_with_content(page_id: usize) -> PageRef {
-        let page = Arc::new(Page::new(page_id));
+        let page = Arc::new(Page::new(page_id as i64));
         {
             let buffer = crate::Buffer::new_temporary(4096);
             let page_content = PageContent {
@@ -1325,7 +1325,7 @@ mod tests {
                     let id_page = rng.next_u64() % max_pages;
                     let key = PageCacheKey::new(id_page as usize);
                     #[allow(clippy::arc_with_non_send_sync)]
-                    let page = Arc::new(Page::new(id_page as usize));
+                    let page = Arc::new(Page::new(id_page as i64));
 
                     if cache.peek(&key, false).is_some() {
                         continue; // Skip duplicate page ids
